LayoutTests/platform/gtk/accessibility/caret-browsing-anchor-followed-expected.txt

A link pointing to an anchor at the end of the document

A paragraph with a some text in the middle

The Anchor

A paragraph with a some text right after the anchor

This tests that the caret position is properly updated when following anchor links.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axLink.role is 'AXRole: link'
PASS axLink.isFocused is true
PASS axLink.isFocused is false
PASS getSelection().anchorNode.nodeValue is 'The Anchor'
PASS getSelection().anchorNode.nodeValue is 'A paragraph with a some text right after the anchor'
PASS successfullyParsed is true

TEST COMPLETE
